wip: move additionalPreviewContent inside popPpPreview content
